software/pupae-thermal-device/pyproject.toml
[project]
name = "pupae-thermal-device"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Will Dickson", email = "wbd@caltech.edu" }
]
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.3",
    "numpy>=2.3.2",
    "pyserial>=3.5",
    "pyside6>=6.9.1",
]

[project.scripts]
pupae-thermal-device-tuner = "pupae_thermal_device:tuner_app_main"

[build-system]
requires = ["uv_build>=0.8.2,<0.9.0"]
build-backend = "uv_build"
